refactor(dir): Rename fixture mod to root
Cherry pick 23c5326b127d06d44c63bec4e909b0fe7f0d53de (#300)
